pyproject.toml
[tool.poetry]
name = "e-commerce-api"
version = "0.0.1"
description = "Desafio Python - E-commerce"
authors = ["Alexandre Lima"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12.3"
django = "^5.1"
djangorestframework = "^3.15.2"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
pytest-coverage = "^0.0"
pytest-django = "^4.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
